new setup

+++ .boot2docker/bootsync.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
 mkdir -p /mnt/sda1/etc/whaler
+if [ ! -L /etc/whaler ]; then
+    ln -s /mnt/sda1/etc/whaler /etc/whaler
+fi
+
 mkdir -p /mnt/sda1/var/lib/whaler
-ln -s /mnt/sda1/etc/whaler /etc/whaler
-ln -s /mnt/sda1/var/lib/whaler /var/lib/whaler
+if [ ! -L /var/lib/whaler ]; then
+    ln -s /mnt/sda1/var/lib/whaler /var/lib/whaler
+fi
